将disablePlugin方法从LiteLoader对象上移除，转到设置界面中
