+++ src/preload.js
@@ -6,7 +6,6 @@ Object.defineProperty(globalThis, "LiteLoader", {
     value: {
         ...ipcRenderer.sendSync("LiteLoader.LiteLoader.LiteLoader"),
         api: {
-            disablePlugin: (...args) => ipcRenderer.invoke("LiteLoader.LiteLoader.api", "disablePlugin", "disablePlugin", ...args),
             config: {
                 get: (...args) => ipcRenderer.invoke("LiteLoader.LiteLoader.api", "config", "get", ...args),
                 set: (...args) => ipcRenderer.invoke("LiteLoader.LiteLoader.api", "config", "set", ...args)
@@ -26,11 +25,11 @@ window.addEventListener("DOMContentLoaded", () => {
     script.type = "module";
     script.defer = true;
     script.src = `local:///${LiteLoader.path.root}/src/renderer.js`;
-    document.head.appendChild(script);
+    document.head.append(script);
 });
 
 
-const preload_codeblock = (code) => `(
+const runPreloadScript = code => binding.createPreloadScript(`(
     async function(
         require,
         process,
@@ -43,7 +42,12 @@ const preload_codeblock = (code) => `(
     ) {
         ${code}
     }
-)`;
+)`)(...arguments);
+
+
+fetch(`local:///${LiteLoader.path.root}/src/setting/preload.js`).then(async res => {
+    runPreloadScript(await res.text());
+});
 
 
 // 加载插件 Preload
@@ -53,7 +57,7 @@ for (const [slug, plugin] of Object.entries(LiteLoader.plugins)) {
     }
     if (plugin.path.injects.preload) {
         fetch(`local:///${plugin.path.injects.preload}`).then(async res => {
-            binding.createPreloadScript(preload_codeblock(await res.text()))(...arguments);
+            runPreloadScript(await res.text());
         });
     }
 }
@@ -63,7 +67,7 @@ for (const [slug, plugin] of Object.entries(LiteLoader.plugins)) {
 let flag = false;
 ipcRenderer.invoke("LiteLoader.LiteLoader.preload").then(preload => {
     if (!flag) {
-        binding.createPreloadScript(preload_codeblock(preload))(...arguments);
+        runPreloadScript(preload);
         flag = true;
     }
 });
